Improved the image upload process when editing the job voucher.
Add the image upload process when editing the job posting to JobItemRepositoryInterface.php and JobItemRepository.php.
The duplicate code of the image upload process when editing the job voucher is summarized in MediaController.php.
Added deletion route of image up screen when editing to sub_image_01.blade.php and sub_image_02.blade.php.